+++ dask_geomodeling/geometry/merge.py
@@ -1,14 +1,6 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-# -*- coding: utf-8 -*-
 """
 Module containing merge operation that act on geometry blocks
 """
-
-from __future__ import print_function
-from __future__ import unicode_literals
-from __future__ import absolute_import
-from __future__ import division
-
 import pandas as pd
 from shapely.geometry import box