+++ dask_geomodeling/measurements.py
@@ -1,5 +1,3 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-# -*- coding: utf-8 -*-
 """
 Module providing a percentile function analogous to the functions provided by
 scipy.ndimage.measurements.