+++ dask_geomodeling/raster/misc.py
@@ -1,5 +1,3 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-# -*- coding: utf-8 -*-
 """
 Module containing miscellaneous raster blocks.
 """