+++ dask_geomodeling/tests/test_raster.py
@@ -1,11 +1,3 @@
-# (c) Nelen & Schuurmans.  GPL licensed, see LICENSE.rst.
-# -*- coding: utf-8 -*-
-
-from __future__ import print_function
-from __future__ import unicode_literals
-from __future__ import absolute_import
-from __future__ import division
-
 from datetime import datetime as Datetime
 from datetime import timedelta as Timedelta
 import unittest